requirements.txt
alabaster==0.7.16
anyio==4.4.0
appnope==0.1.4
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bibtexparser==1.4.1
biopython==1.83
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
docutils==0.20.1
executing==2.0.1
fake-useragent==1.5.1
free-proxy==1.1.1
git-filter-repo==2.38.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imagesize==1.4.1
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
requests==2.32.2
scholarly==1.7.11
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.16.0
wsproto==1.2.0
